requirements.txt
aniso8601==7.0.0
beautifulsoup4==4.8.0
Django==2.2.5
django-bootstrap4==1.0.1
django-graphiql==0.4.4
django-mathfilters==0.4.0
djangorestframework==3.10.3
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
Pillow==6.2.0
promise==2.2.1
psycopg2==2.8.3
python-dotenv==0.10.3
pytz==2019.2
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
whitenoise==4.1.4
